Add usage information in SPARQL to FA script
This patch adds the information on the usage of the converter of SPARQL
queries into a set of adjacency matrices representing NFAs corresponding
to them.

The script itself receives an enumerated directed-labeled graph
represented by adjacency matrix decomposition in MTX format and converts
the provided set of SPARQL queries over the graph into a set of NFAs as
their adjacency matrix decomposition.

In other words, after enumerating the vertices and the labels of an
arbitrary graph in NT format this script allows user to convert SPARQL
queries over the original graph into a set of the queries in the
enumerated graph expressed in terms of linear-algebra.
The suggested conversion pipeline for a set of queries over an
arbitrary graph is:
* Get the knowledge graph.
* Convert the graph into N-Triples format.
* Enumerate the graph using `convert-nt-to-mm.py`.
* Convert the original queries using this script
  (`convert-sparql-to-fa.py`).